dataplane: split the guarded link-map accessors out of loader.go
The #6740 accessor family pushed loader.go from 1413 to 1607 LOC,
crossing the 1500 [WATCH] floor and redding
pkg/refactoraudit.TestTouchedFileCrossedModularityThreshold.

Checked the three docs/refactoring-audit.md "When NOT to refactor"
cases; none applies. loader.go is a large multi-concern file, not one
cohesive definition; nothing here is awaiting replacement; and the split
REDUCES review friction rather than adding it, because the moved block
is a contiguous family this change had just introduced.

What moves is one concern: guarded access to the Manager's
link-membership maps — the snapshot accessors, the per-entry helpers,
the swap-target snapshot, the VLAN writer and the test seam. They share
a single invariant pair (hand out no live map; hold no lock across a
syscall) which is now stated once at the top of the new file instead of
being repeated per method.

Pure relocation: the moved region is BYTE-IDENTICAL to what it replaced
(183 lines, md5 42e1227f2ffd), verified by extracting it from the
pre-move file and diffing against where it now sits. No name changes, so
no call site moved. loader.go ends at 1422 — below the 1413 it started
at, since the extracted block took its `maps` import with it.

gofmt checked on the two touched files only.
